requirements.txt
pdfplumber
pytesseract
Pillow
requests
